requirements.txt
Flask
Flask-SQLAlchemy
APScheduler
speedtest-cli